# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deloitte Initiative Discovery Platform",
    "description": "\n    ## Deloitte Initiative Discovery Platform API\n    \n    A unified platform for Deloitte analysts to discover and engage with firm initiatives.\n    \n    ### Features:\n    * **Authentication**: Azure AD B2C SSO integration\n    * **User Profiles**: Manage analyst and leader profiles\n    * **Initiative Management**: Create, update, and manage initiatives\n    * **Discovery Tools**: Search, filter, and browse initiatives\n    * **AI Recommendations**: Personalized initiative matching\n    * **Engagement**: Save, bookmark, and track initiatives\n    \n    ### User Roles:\n    * **Analyst**: Browse, search, and receive recommendations\n    * **Leader**: Create and manage initiatives\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/search/": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search initiatives",
        "description": "Search and filter initiatives.\n\nSupports:\n- **Text search**: Natural language search across title and description\n- **Skills filter**: Match initiatives requiring specific skills\n- **Practice area**: Filter by practice (Strategy, Technology, etc.)\n- **Industries**: Filter by target industries\n- **Time commitment**: Filter by expected time commitment\n\nExamples:\n- `/search?q=AI healthcare` - Search for AI healthcare initiatives\n- `/search?skills=Python&skills=Machine Learning` - Find initiatives needing Python and ML\n- `/search?practice_area=Technology&time_commitment=5 hours/week` - Technology initiatives with specific commitment\n\nFuture enhancement: This will use vector embeddings for semantic search.",
        "operationId": "search_initiatives_api_v1_search__get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search query (searches title and description)",
              "title": "Q"
            },
            "description": "Search query (searches title and description)"
          },
          {
            "name": "skills",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by required skills",
              "title": "Skills"
            },
            "description": "Filter by required skills"
          },
          {
            "name": "practice_area",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by practice area",
              "title": "Practice Area"
            },
            "description": "Filter by practice area"
          },
          {
            "name": "industries",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by industries",
              "title": "Industries"
            },
            "description": "Filter by industries"
          },
          {
            "name": "time_commitment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by time commitment",
              "title": "Time Commitment"
            },
            "description": "Filter by time commitment"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/semantic": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Semantic search (AI-powered)",
        "description": "AI-powered semantic search using vector embeddings.\n\nExamples:\n- \"Find initiatives involving AI and healthcare\"\n- \"Available short-term volunteering projects\"\n- \"Innovation opportunities in financial services\"\n\nThis endpoint uses vector embeddings to understand the semantic meaning\nof your query and find the most relevant initiatives.\n\n**Note**: This is a placeholder. Full implementation requires:\n1. Embedding generation for initiatives\n2. Vector database integration (Qdrant)\n3. Cosine similarity matching",
        "operationId": "semantic_search_api_v1_search_semantic_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Natural language search query",
              "title": "Query"
            },
            "description": "Natural language search query"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/": {
      "get": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Get personalized recommendations",
        "description": "Get AI-powered personalized initiative recommendations.\n\nRecommendations are based on:\n- **User profile**: Skills, interests, industry preferences\n- **Past interactions**: Previously viewed and saved initiatives\n- **Semantic matching**: Profile and initiative embeddings\n- **Collaborative filtering**: Similar users' preferences\n\nEach recommendation includes:\n- The initiative details\n- Match score (0-1)\n- Explanation of why it was recommended\n\n**Note**: This is a simplified implementation. Full AI recommendation requires:\n1. User profile embeddings\n2. Initiative embeddings\n3. Interaction history analysis\n4. Hybrid recommendation algorithm (content + collaborative filtering)",
        "operationId": "get_recommendations_api_v1_recommendations__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Number of recommendations",
              "default": 10,
              "title": "Limit"
            },
            "description": "Number of recommendations"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecommendationResponse"
                  },
                  "title": "Response Get Recommendations Api V1 Recommendations  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/user/{user_id}": {
      "get": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Get recommendations for user",
        "description": "Get recommendations for a specific user.\n\nThis endpoint allows leaders/admins to see what initiatives\nwould be recommended for specific analysts.",
        "operationId": "get_user_recommendations_api_v1_recommendations_user__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecommendationResponse"
                  },
                  "title": "Response Get User Recommendations Api V1 Recommendations User  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InitiativeDuration": {
        "type": "string",
        "enum": [
          "short_term",
          "ongoing",
          "fixed_duration"
        ],
        "title": "InitiativeDuration"
      },
      "InitiativeList": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InitiativeResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items",
          "page",
          "page_size"
        ],
        "title": "InitiativeList"
      },
      "InitiativeResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "practice_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Practice Area"
          },
          "skills_needed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Needed"
          },
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries"
          },
          "time_commitment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Commitment"
          },
          "duration": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InitiativeDuration"
              }
            ],
            "default": "ongoing"
          },
          "duration_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Details"
          },
          "role_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Type"
          },
          "contact_person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Person"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/InitiativeStatus"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count",
            "default": 0
          },
          "save_count": {
            "type": "integer",
            "title": "Save Count",
            "default": 0
          },
          "application_count": {
            "type": "integer",
            "title": "Application Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "id",
          "status",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "InitiativeResponse"
      },
      "InitiativeStatus": {
        "type": "string",
        "enum": [
          "open",
          "active",
          "full",
          "closed"
        ],
        "title": "InitiativeStatus"
      },
...
      "RecommendationResponse": {
        "properties": {
          "initiative": {
            "$ref": "#/components/schemas/InitiativeResponse"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "initiative",
          "score",
          "explanation"
        ],
        "title": "RecommendationResponse",
        "description": "Recommendation with explanation"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}